+++ src/qdmi/sc/Device.cpp
@@ -157,7 +157,13 @@ auto Device::sessionFree(MQT_SC_QDMI_Device_Session session) -> void {
 }
 auto Device::queryProperty(const QDMI_Device_Property prop, const size_t size,
                            void* value, size_t* sizeRet) const -> int {
-  if ((value != nullptr && size == 0) || prop >= QDMI_DEVICE_PROPERTY_MAX) {
+  if ((value != nullptr && size == 0) ||
+      (prop >= QDMI_DEVICE_PROPERTY_MAX &&
+       prop != QDMI_DEVICE_PROPERTY_CUSTOM1 &&
+       prop != QDMI_DEVICE_PROPERTY_CUSTOM2 &&
+       prop != QDMI_DEVICE_PROPERTY_CUSTOM3 &&
+       prop != QDMI_DEVICE_PROPERTY_CUSTOM4 &&
+       prop != QDMI_DEVICE_PROPERTY_CUSTOM5)) {
     return QDMI_ERROR_INVALIDARGUMENT;
   }
   ADD_STRING_PROPERTY(QDMI_DEVICE_PROPERTY_NAME, name_.c_str(), prop, size,
@@ -209,7 +215,12 @@ auto MQT_SC_QDMI_Device_Session_impl_d::setParameter(
     QDMI_Device_Session_Parameter param, const size_t size,
     const void* value) const -> int {
   if ((value != nullptr && size == 0) ||
-      param >= QDMI_DEVICE_SESSION_PARAMETER_MAX) {
+      (param >= QDMI_DEVICE_SESSION_PARAMETER_MAX &&
+       param != QDMI_DEVICE_SESSION_PARAMETER_CUSTOM1 &&
+       param != QDMI_DEVICE_SESSION_PARAMETER_CUSTOM2 &&
+       param != QDMI_DEVICE_SESSION_PARAMETER_CUSTOM3 &&
+       param != QDMI_DEVICE_SESSION_PARAMETER_CUSTOM4 &&
+       param != QDMI_DEVICE_SESSION_PARAMETER_CUSTOM5)) {
     return QDMI_ERROR_INVALIDARGUMENT;
   }
   if (status_ != Status::ALLOCATED) {
@@ -277,7 +288,12 @@ auto MQT_SC_QDMI_Device_Job_impl_d::setParameter(
     const QDMI_Device_Job_Parameter param, const size_t size, const void* value)
     -> int {
   if ((value != nullptr && size == 0) ||
-      param >= QDMI_DEVICE_JOB_PARAMETER_MAX) {
+      (param >= QDMI_DEVICE_JOB_PARAMETER_MAX &&
+       param != QDMI_DEVICE_JOB_PARAMETER_CUSTOM1 &&
+       param != QDMI_DEVICE_JOB_PARAMETER_CUSTOM2 &&
+       param != QDMI_DEVICE_JOB_PARAMETER_CUSTOM3 &&
+       param != QDMI_DEVICE_JOB_PARAMETER_CUSTOM4 &&
+       param != QDMI_DEVICE_JOB_PARAMETER_CUSTOM5)) {
     return QDMI_ERROR_INVALIDARGUMENT;
   }
   return QDMI_ERROR_NOTSUPPORTED;
@@ -287,7 +303,13 @@ auto MQT_SC_QDMI_Device_Job_impl_d::queryProperty(
     // NOLINTNEXTLINE(readability-non-const-parameter)
     const QDMI_Device_Job_Property prop, const size_t size, void* value,
     [[maybe_unused]] size_t* sizeRet) -> int {
-  if ((value != nullptr && size == 0) || prop >= QDMI_DEVICE_JOB_PROPERTY_MAX) {
+  if ((value != nullptr && size == 0) ||
+      (prop >= QDMI_DEVICE_JOB_PROPERTY_MAX &&
+       prop != QDMI_DEVICE_JOB_PROPERTY_CUSTOM1 &&
+       prop != QDMI_DEVICE_JOB_PROPERTY_CUSTOM2 &&
+       prop != QDMI_DEVICE_JOB_PROPERTY_CUSTOM3 &&
+       prop != QDMI_DEVICE_JOB_PROPERTY_CUSTOM4 &&
+       prop != QDMI_DEVICE_JOB_PROPERTY_CUSTOM5)) {
     return QDMI_ERROR_INVALIDARGUMENT;
   }
   return QDMI_ERROR_NOTSUPPORTED;
@@ -317,7 +339,11 @@ auto MQT_SC_QDMI_Device_Job_impl_d::getResults(
     QDMI_Job_Result result,
     // NOLINTNEXTLINE(readability-non-const-parameter)
     const size_t size, void* data, [[maybe_unused]] size_t* sizeRet) -> int {
-  if ((data != nullptr && size == 0) || result >= QDMI_JOB_RESULT_MAX) {
+  if ((data != nullptr && size == 0) ||
+      (result >= QDMI_JOB_RESULT_MAX && result != QDMI_JOB_RESULT_CUSTOM1 &&
+       result != QDMI_JOB_RESULT_CUSTOM2 && result != QDMI_JOB_RESULT_CUSTOM3 &&
+       result != QDMI_JOB_RESULT_CUSTOM4 &&
+       result != QDMI_JOB_RESULT_CUSTOM5)) {
     return QDMI_ERROR_INVALIDARGUMENT;
   }
   return QDMI_ERROR_NOTSUPPORTED;
@@ -330,7 +356,12 @@ auto MQT_SC_QDMI_Site_impl_d::makeUniqueSite(const uint64_t id)
 auto MQT_SC_QDMI_Site_impl_d::queryProperty(const QDMI_Site_Property prop,
                                             const size_t size, void* value,
                                             size_t* sizeRet) const -> int {
-  if ((value != nullptr && size == 0) || prop >= QDMI_SITE_PROPERTY_MAX) {
+  if ((value != nullptr && size == 0) ||
+      (prop >= QDMI_SITE_PROPERTY_MAX && prop != QDMI_SITE_PROPERTY_CUSTOM1 &&
+       prop != QDMI_SITE_PROPERTY_CUSTOM2 &&
+       prop != QDMI_SITE_PROPERTY_CUSTOM3 &&
+       prop != QDMI_SITE_PROPERTY_CUSTOM4 &&
+       prop != QDMI_SITE_PROPERTY_CUSTOM5)) {
     return QDMI_ERROR_INVALIDARGUMENT;
   }
   ADD_SINGLE_VALUE_PROPERTY(QDMI_SITE_PROPERTY_INDEX, uint64_t, id_, prop, size,
@@ -396,7 +427,13 @@ auto MQT_SC_QDMI_Operation_impl_d::queryProperty(
     size_t* sizeRet) const -> int {
   if ((sites != nullptr && numSites == 0) ||
       (params != nullptr && numParams == 0) ||
-      (value != nullptr && size == 0) || prop >= QDMI_OPERATION_PROPERTY_MAX) {
+      (value != nullptr && size == 0) ||
+      (prop >= QDMI_OPERATION_PROPERTY_MAX &&
+       prop != QDMI_OPERATION_PROPERTY_CUSTOM1 &&
+       prop != QDMI_OPERATION_PROPERTY_CUSTOM2 &&
+       prop != QDMI_OPERATION_PROPERTY_CUSTOM3 &&
+       prop != QDMI_OPERATION_PROPERTY_CUSTOM4 &&
+       prop != QDMI_OPERATION_PROPERTY_CUSTOM5)) {
     return QDMI_ERROR_INVALIDARGUMENT;
   }
   if (sites != nullptr) {